backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
celery==5.5.3
redis==6.4.0
requests
pandas
python-dotenv
colorama
fpdf2
pytest
httpx
PyJWT
